% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bbox.R
\name{create_bbox}
\alias{create_bbox}
\title{Create a Bounding Box (BBox) Polygon}
\usage{
create_bbox(xmin, xmax, ymin, ymax, crs)
}
\arguments{
\item{xmin}{Minimum x-coordinate (longitude).}

\item{xmax}{Maximum x-coordinate (longitude).}

\item{ymin}{Minimum y-coordinate (latitude).}

\item{ymax}{Maximum y-coordinate (latitude).}

\item{crs}{Coordinate Reference System (CRS) to be assigned to the bounding box.}
}
\value{
An `sfc` object representing the bounding box polygon.
}
\description{
This function creates a bounding box polygon based on the given minimum and maximum coordinates.
}
\examples{
Create a bounding box
bbox <- create_bbox(xmin = -180, xmax = 180, ymin = -90, ymax = 90, crs = 4326)
}
