% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_HydroSheed_region.R
\name{determine_hydrosheeds_region}
\alias{determine_hydrosheeds_region}
\title{Determine HydroSHEDS Region}
\usage{
determine_hydrosheeds_region(lon, lat)
}
\arguments{
\item{lon}{Longitude coordinate.}

\item{lat}{Latitude coordinate.}
}
\value{
A character string indicating the region(s) that the coordinates fall into.
}
\description{
This function determines the HydroSHEDS region based on given longitude and latitude.
}
\examples{
region <- determine_hydrosheeds_region(lon = 10, lat = 50)
}
