% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_intersecting_river.R
\name{st_intersecting_river}
\alias{st_intersecting_river}
\title{Intersect Rivers with a Bounding Box}
\usage{
st_intersecting_river(rivers_sf, bbox_polygon)
}
\arguments{
\item{rivers_sf}{An `sf` object containing river data.}

\item{bbox_polygon}{An `sf` or `sfc` object representing the bounding box.}
}
\value{
An `sf` object containing the clipped river segments.
}
\description{
This function clips river segments that intersect with a given bounding box.
}
\examples{
\dontrun{
bbox <- create_bbox(xmin = -10, xmax = 10, ymin = -10, ymax = 10, crs = 4326)
rivers <- download_hydrorivers(region = "eu")
clipped_rivers <- st_intersecting_river(rivers, bbox)
}
}
